requirements.txt
django==3.2.7
gunicorn
django-environ
drf_yasg
djangorestframework
django-cors-headers
psycopg2-binary
